Initial nurikabe model. Visual assets.
Added initial nurikabe game board, which can draw numbers, walls and
islands. This is populated from an ascii grid in the Godot editor. I
experimented with a node-based solution, but a large Nurikabe puzzle
(100x100) requires 10,000+ nodes, and is sluggish in the editor. TileMap
seems like the best compromise between performance and nice visuals.

Added visual assets for tiles, and unused visual assets for a purple
character.

Added asset license.