Improved hidden island logic. ISLAND_MOAT rule.
Before, the 'hidden island' logic would expand an island once, and see
if that blocked in any islands. This had a shortcoming where it only saw
'one space into the future' and would not predict if an island with a
large clue would eventually be blocked.

The new 'hidden island' logic is simpler, and relies on a new
'uncompletable island' algorithm which flood fills each clue, avoiding
other clues. This also organically brought about a new 'island_moat'
rule to cover the unusual cases where cells have to be walls to avoid
blocking an island.