Island bubble edge cases.
The old logic said 'empty areas surrounded by islands must be islands',
but this also includes the entire puzzle. The only reason this didn't
come up is because we always had starting techniques which executed
first. In the absence of those techniques, the island bubble rule would
change the entire board to an island.