Experimenting with new faster solver design.
The old solver iterated over a series of techniques, running every
technique on every cell in the grid. For some techniques which require
flood filling cells in hypothetical scenarios, doing this 200 times for
a 10x20 puzzle was way too slow.

The new solver's goal is to run specific techniques over specific cells.