Added 'island_snug' rule.
This is a new rule specific to the fast solver. The slow solver
made every cell on the board a wall one at a time (n), counted
uncompletable islands for each of those board variants (n^2) resulting
in an O(n^3) slog. However, the slow solver also was unable to
distinguish between those cases, merely calling them all 'island
expansion'

The fast solver instead generates a DFS map (n) and can then calculate
the component size for each clue (1), and if it finds a clue which is in
the perfect size enclosure, it marks all empty cells as islands. This is
O(n) and very fast. I intend to demonstrate how fast with an upcoming
performance test.