DemoFastSolver handles idle steps differently.
Before, it would repeatedly call step, which scheduled tasks when the
queue was empty. But sometimes there would be 200 'check if this cell is
unreachable' tasks in a row, and they would all result in nothing
happening, causing the solver to exit early.

Now, it explicitly schedules tasks when the queue is empty. Sometimes
there are 200 'check if this cell is unreachable' tasks in a row, but
its threshold is much higher. However, it also exits early if no tasks
are scheduled.