Avoid enqueuing ~300 'deduce_empty_square' tasks.
The old logic used to enqueue a reachability check for every square.
These checks were fast, but required a lot of overhead just as far as
creating a task, adding it to the queue, popping it from the queue, etc.

The new logic performs the reachability check during the enqueue step,
only adding a task if the square isn't reachability.

Improves Puzzle #41 performance by ~20% (430 ms -> 350 ms)