Added wall chokepoint logic.
Detecting clue chokepoints required our chokepoint map to answer
questions about size. Detecting wall checkpoints required our chokepoint
map to answer questions about walls. That information is embedded in the
basic chokepoint map as 'special cells'. This is generic, so it could
potentially be used in the future to answer questions like 'does this
branch of the DFS tree contain a clue cell' or something like that.