Bugs: start time NPE, chokepoint performance
Fixed error when bifurcating without calling schedule_tasks to schedule the
bifurcation tasks. This occured during gut tests.

Fixed unnecessary BFS recursion in chokepoint map. The chokepoint map
was traversing the entire puzzle, essentially deducing "There's a
chokepoint across the board for this '2' clue... but it's so far away,
that the 2 can be fulfilled, even if the chokepoint is set." It now,
instad, stops traversing the tree early and does not treat it as a
chokepoint.

This chokepoint tweak improves Nikoli #71 performance by about 30% (13060 ms ->
9730 ms)