Added long island deduction. Bumps performance by 40%.
This new 'long island deduction' is featured in Nikoli #71, and occurs
when a clue is really far away from an unclued island. If it's the only
clue that can reach the island, and it's in a straight
vertical/horizontal line, we can deduce that all the cells in that line
are islands.

This improves Nikoli #71 performance by about 40% (10000 ms -> 7300 ms)

Surprisingly, we still need to bifurcate the same amount for Nikoli #71,
we just bifurcate later.

before: stops=13, scenarios=306, duration=10186.727
after: stops=13, scenarios=254, duration=7276.637