FastDeduction enums. Tests pass in strings.
Added enum for FastDeduction reason, and Vector2i for reasoned cells.
Some deductions like the island_chokepoint result in changes which are
very far from the reasoned island. However, we still want to be able to
follow up with those islands. Having the deductions point to the cells
which caused the deduction is helpful for this.

Tests now pass in strings. Constructing them with enums was going to be
too verbose, but constructing them in tests with strings is easier
anyways.