Expand repeatedly with SqueezeFill.
The solver had some times where it repeatedly expanded the same clue or wall.
For example, it might push a clue out close to a corner, and then expand the
wall toward the corner, and then expand the wall into the corner, and then
expand the wall out of the corner.

I've repurposed Border Hug's "repeatedly flood fill as long as there's only
one option and no matching cells" logic as SqueezeFill. This lets us
batch these changes and rebuild the board fewer times.