Bifurcation tweaks: border walls, squeeze fill
The border wall logic used to be an unusual 'semi bifurcation' which
just ran a squeeze fill and didn't really bifurcate. This is an
interesting idea, but it should also apply to other bifurcations.

I've rewritten the existing 'wall strangle' bifurcation to encompass the
border wall bifurcation. I've also rewritten it and other bifurcations
to leverage the squeeze fill logic.