Added local validation logic which runs on a part of the grid.
This local validation logic runs a simplified subset of rules. For
example, it doesn't check whether all of the board's walls can connect
into a single wall -- it just checks whether any nearby walls are fully
sealed in.

This is about 8x as fast (3.617 -> 0.452). For the 16 bifurcations
performed in puzzles 61-71, 14 of them could detect contradictions with
this faster approach. 1 required the full board "simple" logic, and 1
required the full board "complex" logic.