Removed SolverBoard.get_heat(), Probe.related_cells.
This was an interesting idea where deductions near busy parts of the
board would be prioritized over deductions near idle parts of the board.
But the idea of 'near busy parts of the board' requires CPU cycles to
track and compare, and in the end it didn't make things faster.