Solver wall expansion deduction is more lenient.
Technically, you can only perform a wall expansion if there are two
walls. But practically speaking, all nurikabe puzzles are about 50%
walls so if you have a small wall that makes up 10% of the puzzle, it
definitely needs to expand. This allows the generator to run more
smoothly.

Fixed tests.