Generator: Undo, logging, checkpoints
Generator supports undo. Will retry from a checkpoint 5 times, and then
load from an earlier checkpoint.

Generator stores checkpoints. Checkpoints are saved every time the
generator steps without introducing validation errors.

Generator maintains an event log to support the demo. The demo used to
duplicate the logic of 'keep stepping the generator.... but add a few
log messages'. This is impractical now that the generator's execution
flow is more complex, a log is a better fit. We should eventually roll
this out to the solver as well.