Added island chain logic.
Chains are groups of diagonally connected islands, including connections
to the edge of the board. If a cell would create a cycle within a chain,
and the cycle includes two islands belonging two two different clues,
that cell must be a wall.

I've disabled the 'wall chokepoint' deduction which is redundant with
the chain logic.